in a few of the images there are errors. i don't know why. perhaps the camera data stream is broken?
allocating image 390x390
allocating 8655360 bytes (0.00806093 Gbytes) for 80x84x161 complex float volume
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.
fread read 152099 elements which is not the expected 152100.

real    0m1.012s
user    0m3.044s
sys     0m0.595s